initial commit
I needed the ttutil build system to automate release of the swap 4.2 code mostly for the convenience of the pyswap python package. In that package, I am changing the distribution method of the swap and ttutil code, where each should have their own place and clear information about the license and terms of use.